pyproject.toml
[project]
name = "kuri"
version = "0.1.0"
description = "End-to-end ML system for Indian equity markets. Phase 1: data pipeline."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Rishi" }]
license = { text = "Proprietary" }

dependencies = [
    "polars>=1.18.0",
    "pyarrow>=18.0.0",
    "duckdb>=1.1.0",
    "yfinance>=0.2.50",
    "prefect>=3.1.0",
    "typer>=0.15.0",
    "pydantic>=2.10.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.2",
]

[project.scripts]
kuri = "trading.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.12",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trading"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "UP",   # pyupgrade
    "SIM",  # simplify
    "RUF",  # ruff-specific
    "N",    # pep8-naming
    "C4",   # comprehensions
]
ignore = [
    "E501",  # handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B017", "SIM117"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
pretty = true
files = ["src", "tests"]

[[tool.mypy.overrides]]
module = ["yfinance.*", "duckdb.*", "pandas.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false
# Polars Series methods (.min/.max/.item) return untyped values; comparing
# them with floats in test assertions trips strict mypy without aiding
# correctness. Tests are the right place to relax these.
disable_error_code = ["operator", "union-attr"]

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
filterwarnings = [
    "ignore::DeprecationWarning:yfinance.*",
]
